safari-534.53-branch/LayoutTests/platform/qt/fast/spatial-navigation/snav-ensure-scrolls-with-scrollbars-disabled-by-webkit-expected.txt

Start
End
PASS true is true
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS true is true
